Patch from Tero Pelander <tpeland@tkukoulu.fi>:
Version: pam_ldap-108

Problem:
If you use "nss_base_passwd" that also defines a filter the "scope" part
isn't recogniced.
# Syntax: nss_base_XXX          base?scope?filter

Quick solution:
Here is a quick patch for it but you might want to make a real one that
checks the whole value of "scope" part and not just the first 3 or 4
characters.